# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PartyVisibility = typeof PartyVisibility[keyof typeof PartyVisibility];


export const PartyVisibility = {
  public: 'public',
  private: 'private',
} as const;

export interface Party {
  id: number;
  name: string;
  subject: string;
  description: string;
  host: string;
  accent: string;
  visibility: PartyVisibility;
  memberCount: number;
  maxMembers: number;
  isLive: boolean;
  isPremium: boolean;
  createdAt: string;
}

export type PartyInputVisibility = typeof PartyInputVisibility[keyof typeof PartyInputVisibility];


export const PartyInputVisibility = {
  public: 'public',
  private: 'private',
} as const;

export interface PartyInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  subject: string;
  /** @minLength 2 */
  description: string;
  /** @minLength 1 */
  host: string;
  /**
     * @minimum 2
     * @maximum 50
     */
  maxMembers?: number;
  accent?: string;
  isPremium?: boolean;
  visibility?: PartyInputVisibility;
}

export interface JoinPartyInput {
  /** @minLength 1 */
  memberName: string;
}

export interface PartyMessage {
  id: number;
  partyId: number;
  sender: string;
  content: string;
  createdAt: string;
  /** Set when the message has been deleted. */
  deletedAt?: string | null;
  /** Only present in the response right after you create a message. Save it locally; needed to delete that message later. */
  deleteToken?: string;
}

export interface PartyMessageInput {
  /** @minLength 1 */
  sender: string;
  /** @minLength 1 */
  content: string;
  /** A random id the client generates and persists locally, used only to apply the short cooldown after a flagged message. */
  clientId?: string;
}

export interface Dashboard {
  partyCount: number;
  liveCount: number;
  totalMembers: number;
  messagesSent: number;
  recentParties: Party[];
}

export interface AssistantChatInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  assistantName: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
  context?: string;
}

export interface AssistantChatResponse {
  assistantName: string;
  reply: string;
}