# lib/api-zod/src/generated/types/assistantChatInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantChatInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  assistantName: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
  context?: string;
}