# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Party } from './party';

export interface Dashboard {
  partyCount: number;
  liveCount: number;
  totalMembers: number;
  messagesSent: number;
  recentParties: Party[];
}