# lib/api-zod/src/generated/types/party.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PartyVisibility } from './partyVisibility';

export interface Party {
  id: number;
  name: string;
  subject: string;
  description: string;
  host: string;
  accent: string;
  visibility: PartyVisibility;
  memberCount: number;
  maxMembers: number;
  isLive: boolean;
  isPremium: boolean;
  createdAt: string;
}