# lib/api-zod/src/generated/types/partyInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PartyInputVisibility } from './partyInputVisibility';

export interface PartyInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  subject: string;
  /** @minLength 2 */
  description: string;
  /** @minLength 1 */
  host: string;
  /**
     * @minimum 2
     * @maximum 50
     */
  maxMembers?: number;
  accent?: string;
  isPremium?: boolean;
  visibility?: PartyInputVisibility;
}