# lib/api-zod/src/generated/types/partyMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PartyMessage {
  id: number;
  partyId: number;
  sender: string;
  content: string;
  createdAt: string;
  deletedAt?: string | null;
  deleteToken?: string;
}