# lib/api-zod/src/generated/types/partyMessageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PartyMessageInput {
  /** @minLength 1 */
  sender: string;
  /** @minLength 1 */
  content: string;
  clientId?: string;
}